Add #to_msgpack_obj method
This method is tried first (before #to_msgpack), and the returned
object is not treated as a raw string, but is converted to msgpack
format.

In the example below, both A and B produce the same result except that A
is much more efficient because it does not construct an intermediate
Ruby string.

  class A
    def to_msgpack_obj
      [@a, @b]
    end
  end

  class B
    def to_msgpack
      [@a, @b].to_msgpack
    end
  end